Se agrega mensaje para confirmar eliminacion

+++ src/app/components/product-list/product-list.component.ts
@@ -1,6 +1,8 @@
 import {Component, OnInit} from '@angular/core';
 import {Product} from "../../common/product";
 import {ProductService} from "../../services/product.service";
+import Swal from "sweetalert2";
+
 
 @Component({
   selector: 'app-product-list',
@@ -26,9 +28,30 @@ export class ProductListComponent implements OnInit {
   }
 
   deleteProductById(id: number) {
-    this.productService.deleteProductById(id).subscribe(
-      ()=> this.listProducts()
-    )
+
+    Swal.fire({
+      title: "Estas seguro de eliminar este registro",
+      text: "Una vez ejecutada esta accion no se puede revertir!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Eliminar!",
+      cancelButtonText: "Cancelar"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.productService.deleteProductById(id).subscribe(
+          ()=> this.listProducts()
+        );
+        Swal.fire({
+          title: "Productos!",
+          text: "Producto eliminado correctamente",
+          icon: "success"
+        });
+      }
+    });
+
+
   }
 
 }
